Opal compatibility + Oga support
Opal is a Ruby-to-JavaScript transpiler that involves certain semantics that the
compiled code has to follow. Most importantly, those are as follows:

* No C library extensions
* Assumption that strings are immutable
* Correct knowledge about required files before compilation takes place

Fixing those issues in code will allow us to compile Plurimath to JavaScript, so
that the code will be usable directly in web browsers and environments like
NodeJS. This is the work that we are doing in the `plurimath-js` repository
and therefore any issues related to the JavaScript version of the library should
be directed there.

One of the most important parts is that the Ox library, as of now, requires
a C extension. An approach we took was to port another library that does mostly
the same, namely Oga and then to port Plurimath to support both Ox and Oga.
We also chose Oga, because of apparent support of Oga, along with Nokogiri, in
the `equivalent-xml` library, which we use for comparing XML trees.

Support for both has been implemented by doing a façade pattern named
`Plurimath.xml_engine` that uses Ox directly if needed and Oga via an API that
is mostly compatible with Ox (as per usage patterns in Plurimath codebase).

The Ruby runtime still uses Ox, except when an environment variable
`PLURIMATH_OGA` is set, in which case, we use Oga. We use this implementation
internally to test for Oga shim completeness while still running code on MRI.
If problems arise, this means that it's not Opal fault directly, that problems
happen, but Oga shim issues.

Other changes most likely refer to Opal build system, minor compatibility fixes
(like assumptions that strings are mutable), corrections of paths. All of those
changes should keep the logic mostly the same when running in MRI.

This is not a full solution for Opal compatibility yet. There are still important
parts of that work in progress. It's not expected at this point that either all
Opal or Oga tests will pass. It is expected though that all the tests that are
currently being ran will pass.

Ref: plurimath-js/plurimath-js#3